+++ src/main/java/net/atlas/combatify/util/CurrentShieldBlockingType.java
@@ -24,8 +24,14 @@ import java.util.function.Consumer;
 import static net.atlas.combatify.util.MethodHandler.arrowDisable;
 
 public class CurrentShieldBlockingType extends ShieldBlockingType {
-	public CurrentShieldBlockingType(String name) {
-		super(name);
+
+	public CurrentShieldBlockingType(String name, boolean crouchable, boolean blockHit, boolean canDisable, boolean needsFullCharge, boolean defaultKbMechanics, boolean hasDelay) {
+		super(name, crouchable, blockHit, canDisable, needsFullCharge, defaultKbMechanics, hasDelay);
+	}
+
+	@Override
+	public Factory<? extends BlockingType> factory() {
+		return Combatify.CURRENT_SHIELD_BLOCKING_TYPE_FACTORY;
 	}
 
 	@Override