+++ src/main/java/net/atlas/combatify/util/EmptyBlockingType.java
@@ -18,7 +18,12 @@ import java.util.function.Consumer;
 
 public class EmptyBlockingType extends BlockingType {
 	public EmptyBlockingType(String name) {
-		super(name);
+		super(name, false, false, false, false, false, false);
+	}
+
+	@Override
+	public Factory<? extends BlockingType> factory() {
+		return (name, crouchable, blockHit, canDisable, needsFullCharge, defaultKbMechanics, hasDelay) -> new EmptyBlockingType(name);
 	}
 
 	@Override