+++ src/main/java/net/atlas/combatify/util/NonBannerShieldBlockingType.java
@@ -1,12 +1,20 @@
 package net.atlas.combatify.util;
 
+import net.atlas.combatify.Combatify;
 import net.atlas.combatify.config.ConfigurableItemData;
 import net.minecraft.world.item.ItemStack;
 
 public class NonBannerShieldBlockingType extends ShieldBlockingType {
-	public NonBannerShieldBlockingType(String name) {
-		super(name);
+
+	public NonBannerShieldBlockingType(String name, boolean crouchable, boolean blockHit, boolean canDisable, boolean needsFullCharge, boolean defaultKbMechanics, boolean hasDelay) {
+		super(name, crouchable, blockHit, canDisable, needsFullCharge, defaultKbMechanics, hasDelay);
 	}
+
+	@Override
+	public Factory<? extends BlockingType> factory() {
+		return Combatify.NON_BANNER_SHIELD_BLOCKING_TYPE_FACTORY;
+	}
+
 	@Override
 	public float getShieldBlockDamageValue(ItemStack stack) {
 		ConfigurableItemData configurableItemData = MethodHandler.forItem(stack.getItem());