+++ src/main/java/net/atlas/combatify/util/SwordBlockingType.java
@@ -21,8 +21,14 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class SwordBlockingType extends PercentageBlockingType {
-	public SwordBlockingType(String name) {
-		super(name);
+
+	public SwordBlockingType(String name, boolean crouchable, boolean blockHit, boolean canDisable, boolean needsFullCharge, boolean defaultKbMechanics, boolean hasDelay) {
+		super(name, crouchable, blockHit, canDisable, needsFullCharge, defaultKbMechanics, hasDelay);
+	}
+
+	@Override
+	public Factory<? extends BlockingType> factory() {
+		return Combatify.SWORD_BLOCKING_TYPE_FACTORY;
 	}
 
 	@Override
@@ -79,5 +85,15 @@ public class SwordBlockingType extends PercentageBlockingType {
 	@Override
 	public Component getStrengthTranslationKey() {
 		return Component.translatable("attribute.name.generic.damage_reduction");
+	}
+
+	@Override
+	public boolean isToolBlocker() {
+		return true;
+	}
+
+	@Override
+	public boolean requiresSwordBlocking() {
+		return true;
 	}
 }